eliminacion vaso en cola
